Reliability Improved

+++ include/ChatApp/Net/EPacket.h
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 namespace ChatApp {
-	enum class EPacket : std::uint32_t {
+	enum class EPacket : std::uint8_t {
 		NIL = -1,
 		LOGIN,
 		SV_MESSAGE,
